+++ RundeckAutomation/Functions/Export-RundeckJob.ps1
@@ -5,7 +5,7 @@ function Export-RundeckJob
     Exports a job from Rundeck to an XML file.
 
     .DESCRIPTION
-    The `Export-RundeckJob` function exports a job from Rundeck to an XML file.
+    The `Export-RundeckJob` function exports a job from Rundeck to an XML file.  WARNING: Only works with the API Token and not username and password.
 
     .EXAMPLE
     Export-RundeckJob -Path '.\test.xml' -ProjectName 'demo' -ID 'b090d4c8-585c-4330-8bc6-ad4783089dfd'
@@ -32,10 +32,9 @@ function Export-RundeckJob
     process
     {
 
-        $jobExport = Invoke-RundeckRestMethod -Method 'GET' -ResourcePath "project/$($ProjectName)/jobs/export" -QueryString "idlist=$($ID)"
+        $jobExport = Invoke-RundeckRestMethod -ContentIsXML -ErrorAction 'Stop' -Method 'GET' -ResourcePath "project/$($ProjectName)/jobs/export" -QueryString "idlist=$($ID)"
 
-        Set-Content -Value $jobExport -Path $Path -Force -Encoding UTF8
+        Set-Content -Value $jobExport.InnerXml -Path $Path -Force -Encoding UTF8
 
-        return $jobExport
     }
 }