+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 0.1.1
+- Add support for API token authentication
+- Fix the Export-RundeckJob cmdlet to have correct content type
+
 # 0.1.0
 - Initial version with basic Rundeck cmdlets